Increase token slack
Currently the SDK renews the access token when it will expire in five
seconds. That is two short, and in some situations it means that the
server receiving the token will consider it already expired, just
because it takes more than five seconds for the token to travel from the
client to the server. To reduce the chances of that problem this patch
changes the SDK so that it will renew the access and refresh tokens when
they expire in less than one minute.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>