Add Prometheus metrics
This patch changes the SDK so that it optionally reports the following
metrics using the Prometheus client library:

- `api_outbound_request_count` - Number of API requests sent.
- `api_outbound_request_duration_sum` - Total time to send API requests, in seconds.
- `api_outbound_request_duration_count` - Total number of API requests measured.
- `api_outbound_request_duration_bucket` - Number of API requests organized in buckets.
- `api_outbound_token_request_count` - Number of token requests sent.
- `api_outbound_token_request_duration_sum` - Total time to send token requests, in seconds.
- `api_outbound_token_request_duration_count` - Total number of token requests measured.
- `api_outbound_token_request_duration_bucket` - Number of token requests organized in buckets.

The duration buckets metrics contain an `le` label that indicates the
upper bound. For example if the `le` label is `1` then the value will be
the number of requests that were processed in less than one second.

The API request metrics have the following labels:

- `method` - Name of the HTTP method, for example GET or POST.
- `path` - Request path, for example /api/clusters_mgmt/v1/clusters.
- `code` - HTTP response code, for example 200 or 500.

To calculate the average request duration during the last 10 minutes,
for example, use a Prometheus expression like this:

```
rate(api_outbound_request_duration_sum[10m]) / rate(api_outbound_request_duration_count[10m])
```

In order to reduce the cardinality of the metrics the path label is
modified to remove the identifiers of the objects. For example, if the
original path is `.../clusters/123` then it will be replaced by
`.../clusters/-`, and the values will be accumulated. The line returned
by the metrics server will be like this:

```
api_outbound_request_count{code="200",method="GET",path="/api/clusters_mgmt/v1/clusters/-"} 56
```

The meaning of that is that there were a total of 56 requests to get
specific clusters, independently of the specific identifier of the
cluster.

The token request metrics will contain the following labels:

- `code` - HTTP response code, for example 200 or 500.

The value of the `code` label will be zero when sending the request
failed without a response code, for example if it wasn't possible to
open the connection, or if there was a timeout waiting for the response.

Note that setting this attribute is not enougth to have metrics
published, you also need to create and start a metrics server, as
described in the documentation of the Prometheus library.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>