Fetch tags of model and metamodel
Currently the `model` and `metamodel` targets of the _Makefile_ clone
the corresponding repositories only if needed, and then they always do a
`git fetch` to get the latest code. But they don't use the `--tags`
options and as a result some tags may be missing. This patch adds that
`--tags` option.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>