Use about to expire access token if needed
Currently the SDK always tries to refresh the access token when it
expires in less than one minute. But this fails when there is no refresh
token or credentials that can be used to obtain a new one. This patch
changes it so that in that case it uses that access token, even if it is
about to expire.

Related: https://jira.coreos.com/browse/SDA-1402
Signed-off-by: Juan Hernandez <jhernand@redhat.com>