Update to metamodel 0.0.7
This patch updates the SDK to use version 0.0.7 of the metamodel. The
moree relevant changes in that version of the metamodel are the
following:

- Keep concepts sorted by name.
- Don't generate empty `const` block for errors.
- Add `Copy` method to builders.

The change to keep concepts sorted by name is specially relevant,
because it forces change in almost all the generated files. The reason
is that the names of attributes are now alphabetically sorted.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>