Don't check kinds of add-ons installations
Currently the server isn't returning correctly the `kind` attribute of
add-on installations. This affects clusters as even if they don't have
add-ons installed because each cluster contains a link to the collection
of add-on installations.

This patch chages the SDK so that it will ignore that `kind` attribute.
It will be reverted when the server is fixed.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>