Add cookie support
This patch changes the SDK so that it honors the `Set-Cookie` headers
sent by the server. That itself is quite simple, but it requires
important changes in the code that sends requests to the log so that the
`Cookie` header is reported correctly. The reason is that the `Cookie`
header is built internally by the `Do` method of the HTTP client. In
order to send them to the log it is necessary to move the code that
dumps requests to a round tripper that is called *after* the `Do` method
has added them.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>