Request new tokens when refresh grant retursn `invalid_grant`
There are some situations where the SSO service returns error code
`invalid_grant` and error description `Session not active` when the
refresh token has not yet expired. It is unclear why this happens, but
we need to handle it in order to recover from those situations. This
patch changes the SDK so that in those cases it will request new tokens.

Signed-off-by: Juan Hernandez <jhernand@redhat.com>