Honour tokens context deadline
Currently the `TokensContext` method uses a back-off strategy to retry failed
requests, but it ignores the deadline of the passed context. This patch fixes
that.

One interesting side effect of that is that it is possible to reduce the time
that tests take to execute by explicitly setting the timeout instead of letting
them run up to the default 15 seconds. This reduces the test time from 45
seconds to 5 seconds.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>