Add support for alternative URLs
This patch adds a new `AlternativeURL` method to the connection builder. It can
be used to configure alternative base URLs for certain path prefixes. For
example, to configure the connection to send all requests starting with
`/api/clusters_mgmt` to `my.server.com` instead of `api.openshift.com`:

....
connection, err := sdk.NewConnectionBuilder().
        AlternativeURL("/api/clusters_mgmt", "https://my.server.com").
        Build()
....

Requests not matching the prefix specified in the `AlternativeURL` method will
still be sent to the default URL.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>