Wrap errors
This patch changes the SDK so that it wraps errors using the `%w` format verb.
This way it will then be possible for the user to check the types of the errors.
For example, a user interested in checkinf if an error was caused by a timeout
it can check it like this:

  response, err := connection.ClusersMgmg().V1().Clusters().Get().Send()
  if errors.Is(err, context.DeadlineExceeded) {
         ...
  }

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>