Add support for loading trusted CA files
This patch adds a new `TrustedCAFile` method to the connection builder that can
be used to specify files containing the certificates of the CAs trusted by the
connection.

In order to test that the patch also changes all the tests so that they use
HTTPS instead of HTTP and so that they explicitly specify the CA certificates
to trust.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>