Load string if it doesn't look like a file
This patch changes the `Load` method of the configuration object so that it
will interpret strings that don't end in `.yaml`, `.yml` or `.d` as YAML text.
This is very convenient for embedding the configuration in Go code. For
example:

```go
connection, err := sdk.NewConnectionBuilder().
	Logger(logger).
	Load(`
		url: https://my.server.com
		token_url: https://your.server.com
		tokens:
		- eY...
	`).
	Build()
if err != nil {
	...
}
```

This use case is very common in tests.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>