Add support for customizing error responses
This patch adds to the authentication builder the possibility to configure the
error identifier, error code and operation identifier included in JSON error
responses. For example, the following configuration:

```go
handler, err := authentication.NewHandler().
	Logger(logger).
	KeysFile(keysFile).
	Next(next).
	Service("my_service").
	Error("123").
	OperationID(func(r *http.Request) string {
		return r.Header.Get("X-Operation-ID")
	}).
	Build()
if err != nil {
	...
}
```

Will result in JSON error response bodies like this:

```json
{
  "kind": "Error",
  "id": "123",
  "href": "/api/clusters_mgmt/v1/errors/123",
  "code": "MY-SERVICE-123",
  "reason": "Bearer token is malformed",
  "operation_id": "456"
}
```

Note that in this example the operation identifier is taken from the
`X-Operation-ID` request header.

Resolves: https://github.com/openshift-online/ocm-sdk-go/issues/319
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>