Add metrics package
This patch adds a new `metrics` package that contains an HTTP round
tripper that generates the Prometheus metrics based on a description of
the tree of URL paths of the API. This has the advantage that it
generates the correct metrics even when the application uses the generic
HTTP methods of the connection, like `Get` and `Post`.

Doing this change also requires the version of the metamodel that
generates that tree of URL paths and removes the `X-Metric` header.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>