Add metrics handler wrapper
This patch adds an handler wrapper that creates HTTP handlers that
generate Prometheus metrics. This is intended to replace the current
mechanism used by the servers to generate metrics, which is based on
inspecting the metadata of the mechanism used to implement route
variables. For example, the clusters management service uses the
_gorilla_ package to implement routes, and with each request it inspects
the path pattern in order to remove object identifiers from the path
before adding it to the `path` label of the metrics. This isn't
specially efficient, and it introduces an undesirable dependencly on
that library.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>