Reorder execution of tests
This patch reorders the execution of tests in the Jenkins environment so
that the tests that provide more value (unit tests in this case) are
executed first. This way if those fail the developer will be notified
sooner, and there will be no need to run the other less valuable (and
longer) tests.

The patch also introduces a `tests` Makefile target that does exactly
the same than the existing `test` target. This is because I frequently
type `make tests` and I get upset because it doesn't work.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>