Add _tolerance_ parameter to authentication handler
This patch adds a new _tolerance_ parameter to the authentication
handler. This parameter is the maximum time that tokens can be expired
and anyhow accepted. For example, to accept tokens that have expired up
to ten minutes ago:

```go
handler, err := NewHandler().
       Logger(logger).
       KeysFile(keysFile).
       Tolerance(10 * time.Minute).
       Next(next).
       Build()
if err != nil {
       ...
}
```

The default value for this parameter will be zero tolerance.

Related: https://issues.redhat.com/browse/SDA-4006
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>