Don't require refresh token for client credentials grant
Currently the authentication transport wrapper requires the presence of
the refresh token in all responses from the SSO server. This used to be
true in older versions of the server behind `sso.redhat.com`, but
recently it was changed to not return it for client credentials grant
requests. That means that the SDK will fail to obtain a new token when
the initial one is expired. This patch fixes that issue.

Related: https://issues.redhat.com/browse/SDA-4352
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>