Don't use refresh token if have client credentials
Currently if the SSO server returns a refresh token together with the
access token we will use it to request new tokens. This works fine with
the current SSO server, but could break if the server was more strict
and didn't accept the refresh grant when the session has been initiated
using the client credentials grant. To avoid that potential future
breakage this patch changes the authentication wrapper so that it will
never use the refresh grant if it can use the client credentials grant
instead.

Note that the previous behaviour was slightly more secure: the refresh
token was rotated periodically but the client secret isn't. However the
_OAuth_ standard mandates this approach. See section 4.4.3 of RFC6749.

Related: https://issues.redhat.com/browse/SDA-4352
Related: https://datatracker.ietf.org/doc/html/rfc6749#section-4.4.3
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>