Rename `Random` to `Jitter` and introduce timeouts
This patch renames the `Random` parameter of the leadership flag to
`Jitter` because it reflects better what it is used for.

The patch also introduces timeouts for the database operations so that
the flag will recover faster when database connection packets are
silently dropped.

Note that the `lib/pq` doesn't always respect timeouts, in particular it
doesn't respect them when packags are dropped. To effectively respect
those timeouts it is necessary to switch to the `pgx` family of
libraries. In order to make that easier this patch also adds some
functions to abstract the differences, like the `database.ErrorCode`
that extracts the error code from a database error and works with both
`lib/pq` and `pgx`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>