Retry when server sends go away before settings
This patch adds a test that verifies that the SDK retries operations
when the server incorrectly sends a go away frame before a settings
frame. This seems to be the behavior of the OpenShift router
(implemented with haproxy 2.2.13) when the pod of the service is
restarted.

Related: https://github.com/haproxy/haproxy/issues/13
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>