Fix token close to expire test
This patch checks the test that checks that access tokens are used even
if they are close to expire when there is no other way to get new
tokens. Currently it fails some times because it generates a token that
expires too soon: in one second. It is now changed to 10 seconds.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>