Opaque refresh tokens
Currently we assume that refresh tokens are JSON web tokens, we try to
parse them and we fail if they aren't. But this is a misconception, it
is true for the refresh tokens issued by `sso.redhat.com`, but not true
in general, and not true for _Okta_ in particular. This patch changes
the authentication package so that it will no longer assume that. We
will still try to parse refresh tokens, but will not fail when that
isn't possible.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>