Fix retry wait
Currently the retry interval isn't correctly calculated for the first
retry: it is double what it should be. This isn't detected by tests
because they check only that the retry wait is larger than the minimum
expected value. This patch fixes the issue and improves the tests so
that they use a lower and upper limit.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>