Fix bug - service log is not terminating properly
Modify function Close to return nil in case the connection is already closed,
to avoid causing an error when trying to destroy OSL client when the connection
is already closed.
Adding a test to close a closed connection returns nil.
After this PR is merged - trying to destroy a component when the connection is closed
won't cause an error.

Solve SDA-4085