Support pull-secret access token as a valid token.
This change allows the use of the pull-secret access token
as a valid means of token authentication. If detected as
one of the supplied tokens, it will be used in preference
to the access and refresh tokens.

The pull-secret access token is of the form:
<UUID>:<Base64d Token> and is an accepted means of
authentication for selective AMS endpoints.