Remove unreliable leadership flag tests
The tests that check that the leadership flag can't be raised
simultaneosly by multiple processes aren't reliable because they count
the number of processes that have the flag raised in a loop, and while
the loop is in iteration for process _n_ the state of the flag can have
changed for process _n-m_. In the GitHub actions environment that
results false negatives too frequently. This patch removes those tests.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>