Update to version 4 of JWT library
This patch updates the code of the project to use version 4 of the
`github.com/jwt-go/jwt` library. The new version is backwards compatible
with version 3, but it is part of the contract of the SDK, so
applications using the SDK will also need to update. The only required
change is to replace import `github.com/jwt-go/jwt` with
`github.com/jwt-go/jwt/v4`.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>