Double period used by leadership flag tests
Currently the tests of the leadership flag check the database every 100
ms. This works fine in development environments, but seens to not be
enough in the GitHub actions, probably because the virtual machines that
run the tests don't have much CPU assigned. The results is that the
tests frequently generate false negatives. To try to avoid that this
patch doubles the interval used by those tests.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>