Install metamodel with `go install`
A previous patch changed the build process so that it downloads the
released `metamodel` binary instead of building it from source.
Unfortunately that was done in a way that only works for Linux, because
it is hardcoded to download the Linux binary. It could be changed to
check the operating system and download the corresponding binary, but
that is cumbersome. Instead of that the metamodel project has been
changed so that starting with version 0.0.49 it can be installed with
`go install`, and this patch changes this project to do that.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>