feat: install and use tools at project scope level
Currently, in the project some tools are used to generate Go
types and the OpenAPI specs from the OCM API model.

This tools are used by some Makefile targets of the project.

However, the project assumes that the tools are installed in the
user's environment. This leads to errors when trying to use
the Makefile targets and the user being confused as to why.

Aside from that, even if the user has those tools in the environment, the
version used by those tools might be different than the version that was
used to generate the Go types and OpenAPI specs in the repository's
content, which might lead to differences in the autogenerated
code with respect to what's commited in the repository and between users.

This set of changes ensures that the tools used by the Makefile are the
ones installed by it. This ensures that anyone that used the project
and generates the code sing the Makefile targets will use the same
version of the tools

The tools are now always installed in the `bin` directory
of the project's path.

Users of the project that want to directly use the installed tools
manually should run them using the binaries installed at project
level to ensure that there are no differences in the tool versions.

This set of changes also ensures that `goimports` is also installed
as a project-level binary it as it is a requirement to run
the `metamodel` tool. If the user didn't have it installed this ended
up in errors.

Finally, it removes the explicit install of the `goimports` binary as
it is now automatically installed when running `make generate`