feat: install and use ginkgo at project level
Currently, in the project the ginkgo cli tool is used by
some Makefile targets of the project and at Go level.

However, the project assumes that the cli tool is installed in the
user's environment. This leads to errors when trying to use
the Makefile targets and the user being confused as to why.

Aside from that, even if the user has the tool in the environment, the
version used by t might be different than the version that is specified
in go.mod, which might lead to errors when running the tool and
different observed behavior between users.

This set of changes ensures that the tools used by the Makefile are the
ones installed by it. This ensures that anyone that uses the project
and runs the ginkgo cli tool using the Makefile targets will use the same
version of the tool.

The tool is now always installed in the `bin` directory
of the project's path.

Users of the project that want to directly use the installed tool
manually should run it using the binary installed at project
level to ensure that there are no differences in the tool version.