chore: update golangci-lint in GH action to 1.61.0
The current version was causing CI issues where the
GitHub runner was stopping due to what we think is excessive
usage of resource. Possibly a memory leak. Updating the
golangci-lint version used by the GitHub action fixes
the issue. We also get the newest changes and fixes by
updating the version.